Integrated repository catalog and tags commands
User credentials are not required if the registry is under user's
subscription. Updated Dockerfile to install docker.

+++ src/command_modules/azure-cli-registry/azure/cli/command_modules/registry/custom.py
@@ -23,7 +23,10 @@ from ._utils import (
 
 from ._format import output_format
 
-def _cr_list(resource_group=None):
+import azure.cli._logging as _logging
+logger = _logging.get_az_logger(__name__)
+
+def cr_list(resource_group=None):
     '''Returns the list of container registries.
     :param str resource_group: The name of resource group
     '''
@@ -32,39 +35,52 @@ def _cr_list(resource_group=None):
     else:
         return arm_get_registries_in_subscription()
 
-def _cr_create(resource_group, registry_name, location, storage_account_name=None):
+def cr_create(resource_group, registry_name, location, storage_account_name=None, deployment_name="Microsoft.Krater"):
     '''Returns the created container registry.
     :param str resource_group: The name of resource group
     :param str registry_name: The name of container registry
     :param str location: The name of location
     :param str storage_account_name: The name of storage account
+    :param str deployment_name: The name of deployment  
     '''
     if storage_account_name:
-        return arm_deploy_template(resource_group, registry_name, location, storage_account_name)
+        return arm_deploy_template(resource_group, registry_name, location, storage_account_name, deployment_name)
     else:
         return get_mgmt_service_client().create(resource_group, registry_name, RegistryParameters(location=location))
 
-def _cr_delete(registry_name):
+def cr_delete(registry_name):
     '''Deletes the container registry that matches the registry name
     :param str registry_name: The name of container registry
     '''
     registry = arm_get_registry_by_name(registry_name)
+    if registry is None:
+        logger.error('No container registry can be found with name: ' + registry_name)
+        raise SystemExit(1)
+
     resource_group = get_resource_group_by_registry(registry)
     return get_mgmt_service_client().delete(resource_group, registry_name)
 
-def _cr_show(registry_name):
+def cr_show(registry_name):
     '''Returns the container registry that matches the registry name.
     :param str registry_name: The name of container registry
     '''
     registry = arm_get_registry_by_name(registry_name)
+    if registry is None:
+        logger.error('No container registry can be found with name: ' + registry_name)
+        raise SystemExit(1)
+
     resource_group = get_resource_group_by_registry(registry)
     return get_mgmt_service_client().get_properties(resource_group, registry_name)
 
-def _cr_update(registry_name, tags=None):
+def cr_update(registry_name, tags=None):
     '''Returns the updated container registry that matches the registry name.
     :param str registry_name: The name of container registry
     '''
     registry = get_registry_by_name(registry_name)
+    if registry is None:
+        logger.error('No container registry can be found with name: ' + registry_name)
+        raise SystemExit(1)
+        
     resource_group = get_resource_group_by_registry(registry)
 
     newTags = registry.tags
@@ -81,8 +97,8 @@ def _cr_update(registry_name, tags=None):
             
     return get_mgmt_service_client().update(resource_group, registry_name, RegistryParameters(location=registry.location, tags=newTags))
 
-cli_command('registry list', _cr_list, simple_output_query=output_format)
-cli_command('registry create', _cr_create, simple_output_query=output_format)
-cli_command('registry delete', _cr_delete, simple_output_query=output_format)
-cli_command('registry show', _cr_show, simple_output_query=output_format)
-cli_command('registry update', _cr_update, simple_output_query=output_format)
+cli_command('registry list', cr_list, simple_output_query=output_format)
+cli_command('registry create', cr_create, simple_output_query=output_format)
+cli_command('registry delete', cr_delete, simple_output_query=output_format)
+cli_command('registry show', cr_show, simple_output_query=output_format)
+cli_command('registry update', cr_update, simple_output_query=output_format)
